Moving test into common one_node_network
